+++ api/controllers/UsersController.js
@@ -4,11 +4,9 @@ const bcrypt = require('bcrypt')
 const jwt = require('jsonwebtoken')
 const { createError } = require('../../error/error')
 
-// db connection
-const db = mongoose.connection
-
 // models
 const User = require('../models/Users')
+const ShoppingCart = require("../models/ShoppingCart")
 
 // create a user
 exports.createUser = (req, res, next) => {
@@ -18,25 +16,35 @@ exports.createUser = (req, res, next) => {
             // get user
             const user = req.body
 
-            // validate user
+            // check if username and password was provided
             if (!user.username) throw createError(400, "username required")
             if (!user.password) throw createError(400, "password required")
+
+            // password length check
+            if (user.password.length < 6) throw createError(400, "password must be at least 6 characters long")
 
             // create new user
             const newUser = await User.create({
                 username: user.username,
                 password: await bcrypt.hash(user.password, 10)
             })
+
+            // save
             await newUser.save()
 
             // response
-            res.status(201).json({ message: "User created" })
+            res.status(201).json(newUser)
 
         } catch (e) {
 
             // duplicate username error
             if (e instanceof mongoose.mongo.MongoServerError && e.code === 11000) {
                 next(createError(400, "username taken"))
+            }
+
+            // validation error
+            if (e.name === "ValidationError") {
+                next(createError(400, "invalid user data"))
             }
 
             next(e)
@@ -48,19 +56,25 @@ exports.createUser = (req, res, next) => {
 exports.deleteUser = (req, res, next) => {
     (async () => {
         try {
+            const userId = req.params.id
+            const tokenUserId = req.userData.id
+
+            // check if token user is param user
+            if (userId !== tokenUserId) throw createError(401, "Not authorized")
+
             // find user
             const user = await User.findOne({
-                _id: req.userData._id
+                _id: userId
             })
 
             // no user found error
             if (!user) throw createError(404, "user not found")
 
-            // delete user
-            await User.deleteOne({
-                _id: req.userData._id
-            })
+            // delete user's shopping cart
+            if (user.shoppingCart) await ShoppingCart.findByIdAndDelete(user.shoppingCart)
 
+            // delete user
+            await user.delete()
 
             // response
             res.status(200).json({ message: "user deleted" })
@@ -99,7 +113,11 @@ exports.login = (req, res, next) => {
             // create json token payload
             const jwtPayLoad = {
                 username: user.username,
-                _id: user._id
+                id: user._id
+            }
+
+            if (user.shoppingCart) {
+                jwtPayLoad.shoppingCartId = user.shoppingCart
             }
 
             // create json token
@@ -117,6 +135,89 @@ exports.login = (req, res, next) => {
 
         } catch (e) {
 
+
+            next(e)
+        }
+    })()
+}
+
+// get user's shopping cart
+exports.getShoppingCart = (req, res, next) => {
+    (async () => {
+        try {
+
+            // get ids
+            const userId = req.params.userId
+            const tokenUserId = req.userData.id
+
+            // user id check
+            if (userId !== tokenUserId) throw createError(401, "Not authorized")
+            
+            // get user document
+            const user = await User.findById(userId)
+            
+            // get user's shopping cart id
+            const shoppingCartId = user.shoppingCart
+            
+            if (!shoppingCartId) throw createError(404, "No shopping cart found")
+            
+            // get shopping cart
+            const shoppingCart = await ShoppingCart.findById(shoppingCartId)
+            
+            res.status(200).json(shoppingCart)
+            
+
+        } catch(e) {
+
+            next(e)
+        }
+    })()
+}
+
+// update a user's shopping cart
+exports.updateShoppingCart = (req, res, next) => {
+    (async () => {
+        try {
+
+            // get ids
+            const userId = req.params.userId
+            const tokenUserId = req.userData.id
+            const newShoppingCart = req.body
+            
+            // user id check
+            if (userId !== tokenUserId) throw createError(401, "Not authorized")
+            
+            // get user document
+            const user = await User.findById(userId)
+            
+            // get user's shopping cart id
+            const shoppingCartId = user.shoppingCart
+            
+            if (!shoppingCartId) throw createError(404, "No shopping cart found")
+            
+            // get shopping cart
+            const shoppingCart = await ShoppingCart.findById(shoppingCartId)
+
+
+             // clear old shopping cart products
+             shoppingCart.products = []
+
+             // update shopping cart
+             for (const newProduct of newShoppingCart.products) {
+                 shoppingCart.products.push({
+                     _id: newProduct._id,
+                     quantity: newProduct.quantity
+                 })
+             }
+ 
+             // save
+             await shoppingCart.save()
+             
+             // response
+             res.status(200).json(shoppingCart)
+            
+
+        } catch(e) {
 
             next(e)
         }