+++ changelog.md
@@ -1,3 +1,11 @@
+# 0.5.0
+
+- Added PATCH /users/{id}/shoppingcart to update a user's shoppingcart
+- Added GET /users/{id}/shoppingcart to get a user's shoppingcart
+- Added GET /shoppingcart/{id} to get a shoppingcart by id
+- Added POST /shoppingcart to create a shopping cart
+- Added PATCH /shoppingcart/{id} to get a shoppingcart by id
+
 # 0.4.0
 
 - Added POST /users endpoint to create users